pyproject.toml
[project]
name = "hapi-project-api"
version = "0.1.0"
description = "an api application"
dependencies = ["starlette>=0.46.2", "uvicorn>=0.34.1", "tortoise-orm>=0.25.0", "dotenv>=0.9.9", "aerich>=0.8.2", "tomlkit>=0.13.2", "asyncpg>=0.30.0", "orjson>=3.10.16", "bcrypt>=4.3.0"]
requires-python = ">=3.13"
readme = "README.md"

[[project.authors]]
name = "Lowell Rueca"
email = "lowellrueca@gmail.com"

[project.license]
text = "MIT"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.scripts]
_.env_file = ".env.dev"

[tool.pdm.scripts.start]
cmd = ["python", "server.py"]

[tool.aerich]
tortoise_orm = "src.infrastructure.db.config"
location = "./migrations"
src_folder = "./."

[dependency-groups]
dev = [
    "faker>=37.1.0",
]
